am26x: MMCSD: fix data type for error status
 -  update cmd and xfer errorStat to 32bit, as
    error status is stored in high 16 bits.

Fixes: MCUSDK-14935

+++ source/drivers/mmcsd/v1/lld/mmcsd_lld.h
@@ -389,9 +389,9 @@ typedef struct {
 typedef struct {
 
 /** Command Error Status */
-    uint16_t                cmdErrorStat;
+    uint32_t                cmdErrorStat;
 /** Transfer Error Status */
-    uint16_t                xferErrorStat;
+    uint32_t                xferErrorStat;
 /** Data buffer index */
     uint8_t                 *dataBufIdx;
 /** Data Block count used for Transfer */
